+++ app/javascript/maps_maplibre/channels/map_channel.js
@@ -10,8 +10,7 @@ export function createMapChannel(options = {}) {
   const { enableLiveMode = false, ...callbacks } = options
   const subscriptions = {
     family: null,
-    points: null,
-    notifications: null
+    points: null
   }
 
   console.log('[MapChannel] Creating channels with enableLiveMode:', enableLiveMode)
@@ -84,30 +83,8 @@ export function createMapChannel(options = {}) {
     console.log('[MapChannel] Live mode disabled, not subscribing to PointsChannel')
   }
 
-  // Subscribe to notifications channel
-  try {
-    subscriptions.notifications = consumer.subscriptions.create('NotificationsChannel', {
-      connected() {
-        console.log('NotificationsChannel connected')
-        callbacks.connected?.('notifications')
-      },
-
-      disconnected() {
-        console.log('NotificationsChannel disconnected')
-        callbacks.disconnected?.('notifications')
-      },
-
-      received(data) {
-        console.log('NotificationsChannel received:', data)
-        callbacks.received?.({
-          type: 'notification',
-          notification: data
-        })
-      }
-    })
-  } catch (error) {
-    console.warn('[MapChannel] Failed to subscribe to notifications channel:', error)
-  }
+  // Note: NotificationsChannel is handled by notifications_controller.js in the navbar
+  // Creating a second subscription here causes issues with ActionCable
 
   return {
     subscriptions,